Move archive's manipulations to dedicated package
The new package implement a cache mechanism. The cache can be refreshed
by checking the indexes' hashes. This means that only the first refresh
is slow, all the subsequent queries are fast.
The cache relies on files stored on disk, which also allow for a quick
restart of the app since the cache is not destroied on every restart.